rename plotter to logger, fix logger node bug with time
