+++ rosbot_controller/launch/folow_path.launch
@@ -1,14 +1,10 @@
 <launch>
 
-    <arg name="traj_type"       default="sin"/>
-    <arg name="move_plan"       default="~"/>
-   
- 
-    <node pkg="rosbot_controller" type="path_publisher.py" name="path_pub" output="screen">
-        <param name="traj_type"     value="$(arg traj_type)" />
-        <param name="move_plan"     value="$(arg move_plan)" />
-    </node>
-
+	<!-- spawn kinematic model -->
+    <node pkg="rosbot_controller" type="model_runner.py" name="model_runner" output="screen">
+        <param name="robot_frame"       value="model_link" />
+        <param name="cmd_topic"         value="/cmd_vel_model" />
+    </node>   
 
     <node pkg="rosbot_controller" type="path_follower.py" name="gazebo_follower" output="screen">
         <param name="robot_frame"       value="base_link" />
@@ -19,10 +15,5 @@
         <param name="robot_frame"       value="model_link" />
         <param name="cmd_topic"         value="/cmd_vel_model" />
     </node>
-
-    <node pkg="rosbot_controller" type="model_runner.py" name="model_runner" output="screen">
-        <param name="robot_frame"       value="model_link" />
-        <param name="cmd_topic"         value="/cmd_vel_model" />
-    </node>    
 
 </launch>
