+++ rosbot_controller/launch/publish_path.launch
@@ -3,7 +3,6 @@
     <arg name="traj_type"       default="sin"/>
     <arg name="move_plan"       default="~"/>
    
- 
     <node pkg="rosbot_controller" type="path_publisher.py" name="path_pub" output="screen">
         <param name="traj_type"     value="$(arg traj_type)" />
         <param name="move_plan"     value="$(arg move_plan)" />
